pyproject.toml
[project]
name = "blog_to_podcast"
version = "0.1.0"
description = "blog_to_podcast using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.193.2,<1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "streamlit>=1.28.0",
    "firecrawl-py>=4.3.6",
    "python-dotenv>=1.0.0",
]

[project.scripts]
blog_to_podcast = "blog_to_podcast.main:run"
blog2podcast = "blog_to_podcast.main:main"
run_crew = "blog_to_podcast.main:run"
train = "blog_to_podcast.main:train"
replay = "blog_to_podcast.main:replay"
test = "blog_to_podcast.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
